requirements.txt
pandas>=1.3.0
requests>=2.26.0
beautifulsoup4>=4.10.0
openpyxl>=3.0.9
google-generativeai>=0.3.0
python-dotenv>=1.0.0